+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## UNRELEASED
+- Adds `Route.getDescription()`
+- Adds option `description` to `Route` constructor
+
 ## v1.1.0
 - Adds better check on `method` option when creating a `Route`
 - Adds `server` as the 4th argument passed to middleware and route handler